+++ resources/views/layout.blade.php
@@ -7,10 +7,43 @@
     <title>Simple blog</title>
 
     <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.3/dist/css/bootstrap.min.css" rel="stylesheet">
+    @vite(['resources/css/app.css', 'resources/js/app.js'])
+
     @yield('styles')
 </head>
 
 <body>
+    <header class="container my-4">
+        @if (Route::has('login'))
+            <nav class="text-end">
+                @auth
+                    <a
+                        href="{{ url('/dashboard') }}"
+                        class="inline-block px-5 py-1.5 dark:text-[#EDEDEC] border-[#19140035] hover:border-[#1915014a] border text-[#1b1b18] dark:border-[#3E3E3A] dark:hover:border-[#62605b] rounded-sm text-sm leading-normal"
+                    >
+                        Dashboard
+                    </a>
+                @else
+                    <a
+                        href="{{ route('login') }}"
+                        class="inline-block px-5 border border-transparent rounded-start-2 py-1 text-decoration-none text-dark"
+                    >
+                        Iniciar sesión
+                    </a>
+
+                    @if (Route::has('register'))
+                        <a
+                            href="{{ route('register') }}"
+                            class="inline-block px-5 border border-transparent rounded-end-2 py-1 text-decoration-none text-dark"
+                        >
+                            Registrarme
+                        </a>
+                    @endif
+                @endauth
+            </nav>
+        @endif
+    </header>
+
     <div class="container">
         @yield('content')
     </div>