+++ resources/views/posts/index.blade.php
@@ -1,25 +1,21 @@
 @extends('layout')
 
 @section('content')
-    <table class="table table-bordered table-striped">
-        <thead class="text-center">
-            <tr>
-                <th>Título</th>
-                <th>Contenido</th>
-                <th>Categoría</th>
-                <th>Autor</th>
-            </tr>
-        </thead>
+    @if (count($posts) > 0)
+        @foreach ($posts as $post)
+            <div class="border border-secondary rounded-2 p-3 mb-3">
+                <div class="fw-bold">{{ $post->title }}</div>
 
-        <tbody>
-            @foreach ($posts as $post)
-            <tr>
-                <td>{{ $post->title }}</td>
-                <td>{{ $post->content }}</td>
-                <td>{{ $post->category_id }}</td>
-                <td>{{ $post->user_id }}</td>
-            </tr>
-            @endforeach
-        </tbody>
-    </table>
+                <p class="mt-1">{{ $post->content }}</p>
+
+                <div class="d-block w-fit badge bg-dark mt-2">{{ $post->category->name }}</div>
+
+                <small class="d-block text-muted mt-3">{{ $post->user->fullname }} | {{ $post->updated_at }}</small>
+            </div>
+        @endforeach
+    @else
+        <div class="alert alert-primary">
+            Aún no existen publicaciones, crea una <a href="{{ route('posts.create') }}" class="text-decoration-underline">aquí</a>.
+        </div>
+    @endif
 @endsection