+++ routes/web.php
@@ -3,9 +3,9 @@
 use Illuminate\Support\Facades\Route;
 use App\Http\Controllers\ProfileController;
 
-Route::get('/', function () {
+/* Route::get('/', function () {
     return view('welcome'); // TODO: Remover ruta cuando se complete la página de posts
-});
+}); */
 
 Route::get('/dashboard', function () {
     return view('dashboard');
@@ -22,7 +22,7 @@ Route::middleware('auth')->group(function () {
 Route::group([
     'namespace' => 'App\Http\Controllers'
 ], function ($route) {
-    // $route->get('/', 'PostController@index');
+    $route->get('/', 'PostController@index');
     $route->resource('posts', 'PostController');
 
     $route->get('categories', 'CategoryController@index');