+++ CHANGELOG.md
@@ -4,3 +4,4 @@ Registro de cambios del proyecto.
 ## [0.0.1] - [Unreleased]
 ### Added
 - Sistema de registro y autenticación de usuarios.
+- Página principal de posts.